ENH: Add SSE41 support to SPECK-128
Add SSE41 support for SPECK-128

+++ randomgen/src/xoroshiro128/xoroshiro128.c
@@ -44,7 +44,7 @@ void xoroshiro128_jump(xoroshiro128_state_t *state)
 
   s0 = 0;
   s1 = 0;
-  for (i = 0; i < sizeof JUMP / sizeof *JUMP; i++)
+  for (i = 0; i < (int)(sizeof(JUMP) / sizeof(*JUMP)); i++)
     for (b = 0; b < 64; b++)
     {
       if (JUMP[i] & UINT64_C(1) << b)
